+++ .dockerignore
@@ -1,92 +1,94 @@
-# Python 相關
-__pycache__/
+# Git
+.git
+.gitignore
+
+# Python
+__pycache__
 *.pyc
 *.pyo
 *.pyd
-*.pyw
 .Python
-pip-log.txt
-pip-delete-this-directory.txt
-.pytest_cache/
-.coverage
-.tox/
-.cache
-nosetests.xml
-coverage.xml
-*.cover
-*.log
+build
+develop-eggs
+dist
+downloads
+eggs
+.eggs
+lib
+lib64
+parts
+sdist
+var
+wheels
+*.egg-info
+.installed.cfg
+*.egg
 
-# 虛擬環境
+# Virtual environments
 .env
 .venv
-env/
-venv/
-ENV/
-env.bak/
-venv.bak/
+env
+venv
+ENV
+env.bak
+venv.bak
 
-# IDE 和編輯器
-.vscode/
-.idea/
+# IDE
+.vscode
+.idea
 *.swp
 *.swo
-*~
 
-# 系統檔案
+# System files
 .DS_Store
-.DS_Store?
-._*
-.Spotlight-V100
-.Trashes
-ehthumbs.db
 Thumbs.db
 
-# Git 相關
-.git/
-.gitignore
-.gitattributes
-
-# 文檔
-README.md
-LICENSE
-docs/
-*.md
-
-# 開發和測試檔案
-tests/
-test_*.py
-*_test.py
+# 測試和開發檔案
+test_images/
+sample_images/
+test_output/
+*.coverage
+htmlcov/
 .pytest_cache/
-test-results/
 
-# 暫存和備份檔案
+# 日誌
+*.log
+logs/
+
+# 臨時檔案
 *.tmp
 *.bak
-*.backup
-*.old
-*.orig
-
-# Docker 相關
-Dockerfile*
-docker-compose*.yml
-.dockerignore
-
-# CI/CD 和部署
-.github/
-.gitlab-ci.yml
-.travis.yml
-Jenkinsfile
-deploy/
-
-# 大型檔案和資料夾 (根據需要調整)
-data/
-datasets/
-logs/
 temp/
 tmp/
 
-# 機密檔案
-.env.*
-config.json
-secrets/
+# 上傳檔案
+uploads/
+static/uploads/
+
+# 快取檔案
+.cache/
+cache/
+
+# Jupyter Notebook
+.ipynb_checkpoints
+
+# Claude Code 相關
+.claude/
+
+# Docker 相關
+Dockerfile.dev
+docker-compose.dev.yml
+docker-compose.override.yml
+
+# 執行時檔案
+*.pid
+*.sock
+
+# 私有檔案
 private/
+secrets/
+
+# 文檔（保留重要的）
+CLAUDE.md
+GITHUB_WORKFLOW.md
+docs/
