+++ requirements.txt
@@ -1,11 +1,31 @@
+# 核心依賴 - 生產環境必需
+fastapi==0.115.13
+uvicorn[standard]==0.34.3
+pydantic==2.11.7
+pydantic_core==2.33.2
+
+# AI 與圖像處理
+onnxruntime==1.22.0
+opencv-python-headless==4.11.0.86
+pillow==11.2.1
+numpy==2.3.0
+
+# 網路與 API
+requests==2.32.4
+python-multipart==0.0.20
+slowapi==0.1.9
+
+# 工具與其他
+python-dotenv==1.1.0
+coloredlogs==15.0.1
+
+# 自動產生的依賴（保持相容性）
 annotated-types==0.7.0
 anyio==4.9.0
 certifi==2025.6.15
 charset-normalizer==3.4.2
 click==8.2.1
-coloredlogs==15.0.1
 Deprecated==1.2.18
-fastapi==0.115.13
 flatbuffers==25.2.10
 h11==0.16.0
 httptools==0.6.4
@@ -13,26 +33,15 @@ humanfriendly==10.0
 idna==3.10
 limits==5.4.0
 mpmath==1.3.0
-numpy==2.3.0
-onnxruntime==1.22.0
-opencv-python-headless==4.11.0.86
 packaging==25.0
-pillow==11.2.1
 protobuf==6.31.1
-pydantic==2.11.7
-pydantic_core==2.33.2
-python-dotenv==1.1.0
-python-multipart==0.0.20
 PyYAML==6.0.2
-requests==2.32.4
-slowapi==0.1.9
 sniffio==1.3.1
 starlette==0.46.2
 sympy==1.14.0
 typing-inspection==0.4.1
 typing_extensions==4.14.0
 urllib3==2.4.0
-uvicorn==0.34.3
 uvloop==0.21.0
 watchfiles==1.1.0
 websockets==15.0.1
